feat: Refactor logging and structure config files
Removed standalone config.py and instead introduced a structured config folder with separate files for constants and logger setup. This makes the configuration more flexible and easier to extend. Notably, print statements were replaced with logging calls to dedicate a central place for logs.

Moreover, the number of requests in client.py was reduced to 10 from 100 for better performance in testing.